# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Employee Database Application API",
      "description": "Find how to use the Application API",
      "license": {
        "name": "Akhil",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "165.227.109.228:3000",
    "basePath": "/api/v1",
    "tags": [
      {
        "name": "Employees",
        "description": "API for employee in the system"
      }
    ],
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/employee": {
        "post": {
          "tags": [
            "Employees"
          ],
          "description": "Create new employee in system",
          "parameters": [
            {
              "name": "employee",
              "in": "body",
              "description": "Employee that we want to create",
              "schema": {
                "$ref": "#/definitions/Employee"
              }
            }
          ],
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Record added to the database succesfully",
              "schema": {
                "$ref": "#/definitions/Employee"
              }
            }
          }
        },
        "put": {
          "summary": "Update employee with give ID",
          "tags": [
            "Employees"
          ],
          "parameters": [
            {
              "name": "employee",
              "in": "body",
              "description": "Employee with new values of properties",
              "schema": {
                "$ref": "#/definitions/Employee"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Employee is updated",
              "schema": {
                "$ref": "#/definitions/Employee"
              }
            }
          }
        },
        "patch": {
          "summary": "Update few properties of the employee with give ID",
          "tags": [
            "Employees"
          ],
          "parameters": [
            {
              "name": "employee",
              "in": "body",
              "description": "Employee with new values of properties",
              "schema": {
                "$ref": "#/definitions/EmployeePatch"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Employee is updated",
              "schema": {
                "$ref": "#/definitions/EmployeePatch"
              }
            }
          }
        },
        "get": {
          "tags": [
            "Employees"
          ],
          "summary": "Get all employees in system",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Employee"
              }
            }
          }
        }
      },
      "/employee/{emp_id}": {
        "parameters": [
          {
            "name": "emp_id",
            "in": "path",
            "required": true,
            "description": "ID of employee that we want to find",
            "type": "integer"
          }
        ],
        "get": {
          "tags": [
            "Employees"
          ],
          "summary": "Get employee with given ID",
          "responses": {
            "200": {
              "description": "Employee found",
              "schema": {
                "$ref": "#/definitions/Employee"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete employee with given ID",
          "tags": [
            "Employees"
          ],
          "responses": {
            "200": {
              "description": "Record deleted succesfully",
              "schema": {
                "$ref": "#/definitions/Employee"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "Employee": {
        "required": [
          "emp_no",
          "email"
        ],
        "properties": {
          "emp_no": {
            "type": "integer",
            "uniqueItems": true
          },
          "extension":{
            "type":"string"
          },
          "email": {
            "type": "string",
            "uniqueItems": true
          },
          "lastName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "officeCode":{
            "type":"string",
            "description":"use from 1,2,3 office codes"
          },
          "reportTo":{
            "type":"integer",
            "description":"use from 1088,1002,1143,1056,1621"
          },
          "jobTitle":{
             "type":"string"
          }
        }
      },
      "EmployeePatch": {
        "required": [
          "emp_no"
        ],
        "properties": {
          "emp_no": {
            "type": "integer",
            "uniqueItems": true
          },
          "reportTo":{
            "type":"integer",
            "description":"use from 1088,1002,1143,1056,1621"
          },
          "jobTitle":{
             "type":"string"
          }
        }
      }
    }
  }